refund-fraud-ai/requirements.txt
tensorflow
opencv-python
pillow
imagehash
numpy
fastapi
uvicorn
